+++ src/code_generator/common/templates/ocpp_types.hpp.jinja
@@ -51,10 +51,6 @@ struct {{type.name}} {
 {% endif %}
 {% endfor %}
 {%- endif %}
-{% if last %}
-{% if namespace == "v201" %}
-    struct RequiredComponentVariable : ComponentVariable {};
-{% endif %}
 } // namespace {{namespace}}
 } // namespace ocpp
